// src/components/types/data.ts
  export interface TayblePropse {
    id: number;
    car: {
      img: string;
      CarName: string;
      BrandName: string;
      Cost: string;
      company: string;
      place: string;
      license: string;
      info: string;
      ID: string;
      Dates: string;
      Viewed: string;
    };
  }
  export interface TayblePropses {
    id: number;
    img: string;
    CarName: string;
    BrandName: string;
    Cost: string;
    company: string;
    place: string;
    license: string;
    info: string;
    ID: string;
    Dates: string;
    Viewed: string;
  }
export interface CampcarMoto {
    id: number;
    img: string;
    CarName: string;
    BrandName: string;
    Cost: number;
    Dates: number; 
    place: string;
    fuelType: string;
    sleepingCapacity: number;
    company: string;
    license: string;
    info: string;
    Viewed: string;
    transmission: string;
    kitchen: boolean;
    bathroom: boolean;
    airConditioning: boolean;
    waterTankCapacity: number;
    solarPanel: boolean;
    gps: boolean;
    parkingSensors: boolean;
    backupCamera: boolean;
    insurance: boolean;
    maintenancePlan: boolean;
    imageUrl: string;
    hasStorage: boolean;
    hasBathroom: boolean;
    hasAirConditioning: boolean;
    fuelTankCapacity: number;
}
export interface CampcarCamp {
  id: number;
  img: string;
  campingName: string;
  location: string;
  number: string;
  sleepingCapacity: number;
  license: string;
  info: string;
  Viewed: string;
  transmission: string;
  kitchen: boolean;
  bathroom: boolean;
  airConditioning: boolean;
  waterTankCapacity: number;
  solarPanel: boolean;
  gps: boolean;
  parkingSensors: boolean;
  backupCamera: boolean;
  insurance: boolean;
  maintenancePlan: boolean;
  fuelTankCapacity: number;
}

  
  

  
  



  
  